requirements.txt
altgraph==0.17.4
android==1.0
anyio==4.6.2.post1
astroid==2.15.8
audiostream==0.2
bidict==0.23.1
blinker==1.9.0
buildozer==1.5.0
certifi==2024.8.30
cffi==1.16.0
click==8.1.7
cmake @ file:///home/n0n3m4/Desktop/Pydroid/zBuildPrepared/cmake-3.26.4-py3-none-any.whl#sha256=695389dbf5c0aa220776821172e2f03865e0fe9ba10d7459373597d21af6643e
dill==0.3.7
distlib==0.3.9
distro==1.8.0
docutils==0.21.2
filelock==3.16.1
Flask==3.1.0
Flask-SocketIO==5.4.1
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
isort==5.12.0
itsdangerous==2.2.0
jedi==0.19.0
Jinja2==3.1.4
Kivy==2.2.1
lazy-object-proxy==0.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
ninja @ file:///home/n0n3m4/Desktop/Pydroid/zBuildPrepared/ninja-1.11.1-py3-none-any.whl#sha256=bde2ef34b4295ff07a0c1703ce5c7f3acccd621a91b21d4c6aaf09bd8f6591a8
numpy==1.26.1
packaging==23.1
parso==0.8.3
passlib==1.7.4
pexpect==4.9.0
pillow==11.0.0
pkgconfig==1.5.5
platformdirs==3.10.0
psutil==5.9.5
ptyprocess==0.7.0
pybind11==2.11.1
pycparser==2.21
pygame==2.5.0
Pygments==2.18.0
pyinstaller==6.11.0
pyinstaller-hooks-contrib==2024.9
pyjnius==1.5.0
pylint==2.17.5
PyOpenGL==3.1.7
PyOpenGL-accelerate==3.1.7
PySDL2==0.9.15
python-engineio==4.10.1
python-socketio==5.11.4
python-telegram-bot==21.7
scapy==2.6.1
scikit-build==0.17.6
sh==2.1.0
simple-websocket==1.1.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
tomlkit==0.12.1
typed-ast==1.5.5
typing_extensions==4.12.2
virtualenv==20.27.1
watchdog==6.0.0
Werkzeug==3.1.3
wrapt==1.15.0
wsproto==1.2.0
